Move CSW plotter into data_visualization module
